Split octave spectrum classes into separate files
This will help migrate to QML
